added aggregate percent for select multiple
aggregate percent for select multiple added. one additional argument of data required in order to link to all columns of the select multiple question.

To be decided: how to deal with "other" responses where text is filled. That they have been manually removed prior to analysis was the assumption in writing this function.